GraphSegmentation/requirements.txt
matplotlib==3.9.2
networkx==3.4.2
numpy==2.1.3
opencv_python==4.10.0.84
Pillow==9.0.1
Pillow==11.0.0
PyMaxflow==1.2.13
scikit_learn==1.5.2
